pyproject.toml
[tool.poetry]
name = "poetry add django"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "n"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = {extras = ["django"], version = "^5.3.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
